requirements.txt
IPython==8.21.0
streamlit==1.31.1
google.generativeai==0.3.2
